An article, including news and scholarly articles.

An article, including news and scholarly articles.

Article(
  about,
  alternateNames,
  authors,
  content,
  dateAccepted,
  dateCreated,
  dateModified,
  datePublished,
  dateReceived,
  description,
  editors,
  fundedBy,
  funders,
  genre,
  id,
  identifiers,
  images,
  isPartOf,
  keywords,
  licenses,
  meta,
  name,
  pageEnd,
  pageStart,
  pagination,
  parts,
  publisher,
  references,
  text,
  title,
  url,
  version
)

Arguments

about
The subject matter of the content.

alternateNames
Alternate names (aliases) for the item.

authors
The authors of this creative work.

content
The structured content of this creative work c.f. property text.

dateAccepted
Date/time of acceptance.

dateCreated
Date/time of creation.

dateModified
Date/time of most recent modification.

datePublished
Date of first publication.

dateReceived
Date/time that work was received.

description
A description of the item.

editors
People who edited the CreativeWork.

fundedBy
Grants that funded the CreativeWork; reverse of fundedItems.

funders
People or organizations that funded the CreativeWork.

genre
Genre of the creative work, broadcast channel or group.

id
The identifier for this item.

identifiers
Any kind of identifier for any kind of Thing.

images
Images of the item.

isPartOf
An item or other CreativeWork that this CreativeWork is a part of.

keywords
Keywords or tags used to describe this content. Multiple entries in a keywords list are typically delimited by commas.

licenses
License documents that applies to this content, typically indicated by URL.

meta
Metadata associated with this item.

name
The name of the item.

pageEnd
The page on which the article ends; for example "138" or "xvi".

pageStart
The page on which the article starts; for example "135" or "xiii".

pagination
Any description of pages that is not separated into pageStart and pageEnd; for example, "1-6, 9, 55".

parts
Elements of the collection which can be a variety of different elements, such as Articles, Datatables, Tables and more.

publisher
A publisher of the CreativeWork.

references
References to other creative works, such as another publication, web page, scholarly article, etc.

text
The textual content of this creative work.

title
The title of the creative work.

url
The URL of the item.

version
The version of the creative work.

See also