A video file.

A video file.

VideoObject(
  contentUrl,
  alternateNames,
  authors,
  bitrate,
  caption,
  content,
  contentSize,
  dateAccepted,
  dateCreated,
  dateModified,
  datePublished,
  dateReceived,
  description,
  editors,
  embedUrl,
  format,
  fundedBy,
  funders,
  id,
  identifiers,
  isPartOf,
  keywords,
  licenses,
  meta,
  name,
  parts,
  publisher,
  references,
  text,
  thumbnail,
  title,
  transcript,
  url,
  version
)

Arguments

contentUrl
URL for the actual bytes of the media object, for example the image file or video file. Required.

alternateNames
Alternate names (aliases) for the item.

authors
The authors of this creative work.

bitrate
Bitrate in megabits per second (Mbit/s, Mb/s, Mbps).

caption
The caption for this video recording.

content
The structured content of this creative work c.f. property text.

contentSize
File size in megabits (Mbit, Mb).

dateAccepted
Date/time of acceptance.

dateCreated
Date/time of creation.

dateModified
Date/time of most recent modification.

datePublished
Date of first publication.

dateReceived
Date/time that work was received.

description
A description of the item.

editors
People who edited the CreativeWork.

embedUrl
URL that can be used to embed the media on a web page via a specific media player.

format
Media type (MIME type) as per http://www.iana.org/assignments/media-types/media-types.xhtml.

fundedBy
Grants that funded the CreativeWork; reverse of fundedItems.

funders
People or organizations that funded the CreativeWork.

id
The identifier for this item.

identifiers
Any kind of identifier for any kind of Thing.

isPartOf
An item or other CreativeWork that this CreativeWork is a part of.

keywords
Keywords or tags used to describe this content. Multiple entries in a keywords list are typically delimited by commas.

licenses
License documents that applies to this content, typically indicated by URL.

meta
Metadata associated with this item.

name
The name of the item.

parts
Elements of the collection which can be a variety of different elements, such as Articles, Datatables, Tables and more.

publisher
A publisher of the CreativeWork.

references
References to other creative works, such as another publication, web page, scholarly article, etc.

text
The textual content of this creative work.

thumbnail
Thumbnail image of this video recording.

title
The title of the creative work.

transcript
The transcript of this video recording.

url
The URL of the item.

version
The version of the creative work.

See also